Fixing integer-related issues on Mac OS X

+++ python-git/python/yt/bindings/buffered_stream.h
@@ -15,9 +15,9 @@ class TBufferedStream
     : public NConcurrency::IAsyncOutputStream
 {
 public:
-    explicit TBufferedStream(i64 bufferSize);
+    explicit TBufferedStream(size_t bufferSize);
 
-    TSharedRef Read(i64 size = 0);
+    TSharedRef Read(size_t size = 0);
 
     bool Empty() const;
 
@@ -28,8 +28,8 @@ public:
     virtual TAsyncError GetReadyEvent() override;
 
 private:
-    i64 Size_;
-    i64 AllowedSize_;
+    size_t Size_;
+    size_t AllowedSize_;
 
     TSharedRef Data_;
     char* Begin_;
@@ -51,7 +51,7 @@ private:
 
     void Reallocate(size_t len);
     void Move(char* dest);
-    TSharedRef ExtractChunk(i64 size);
+    TSharedRef ExtractChunk(size_t size);
 };
 
 ///////////////////////////////////////////////////////////////////////////////
